+++ R/download_rates.R
@@ -335,6 +335,8 @@ FedInvestData <- function() {
 #'     \item \bold{Adj.Close}
 #'     }
 #'
+#' @importFrom utils read.table
+#'
 #' @author George Fisher
 #'
 #' @examples
@@ -373,6 +375,8 @@ SP500 <- function() {
 #'
 #' @author George Fisher
 #'
+#' @importFrom utils read.table
+#'
 #' @examples
 #' sp500_idx_tr <- SP500TR()
 #' head(sp500_idx_tr)
@@ -384,4 +388,237 @@ SP500TR <- function() {
                         header = TRUE,sep=",",stringsAsFactors = FALSE)
     table$Date <- as.Date(table$Date, "%Y-%m-%d")
     return(table)
+}
+
+#' Coerce a slice of FedInvest data to xts
+#'
+#' @description
+#' The FedInvest data contains one row per bond on a particular date.
+#' This function converts that format into an xts frame.
+#'
+#' @details
+#' The FedInvest data has one row per date, per bond. For analyses such as
+#' Nelson Seigel and Svensson we need the data in an xts frame with the format
+#' of one row per date and one column per maturity.
+#'
+#' This function performs this primary transformation: The bond data has a
+#' continuous set of maturities; a fixed xts table needs discrete maturities.
+#' So, for example, the 1-year maturity data comprises the average YTM for all
+#' the bonds with maturities between 0.5 and 1.5 years.
+#'
+#' A number of minor transformations are also performed such as converting
+#' the RATE data from a character percent format (7.82\%) to decimal (0.0782)
+#'
+#' @importFrom magrittr %>%
+#' @importFrom stats complete.cases
+#'
+#' @author George Fisher
+#'
+#' @param bond_data a slice of the FedInvest data
+#' @return a list
+#' \itemize{
+#'      \item \bold{$rate_table_xts} the xts frame
+#'      \item \bold{$unique_maturities} a numeric vector of the maturities
+#'      for which there is a column in the xts frame in fractional-year format
+#'      (1-month = 1/12 = 0.08333333). It should be noted that if bonds in a
+#'      matruity range of say 15.5-16.5 are not found in the data, there will
+#'      not be a 16.00000 maturity column in the xts frame.
+#'      \item \bold{$modified_data} the original data with additional columns
+#'          \itemize{
+#'              \item \bold{fracyrs} fractional-year time to maturity
+#'              \item \bold{maturities} the regularized fractional-year time to maturity
+#'              \item \bold{coupon} numeric coupon from percent-format RATE
+#'              \item \bold{mid} the mid-point between BUY & SELL
+#'              \item \bold{YTM} yield to maturity
+#'          }
+#' }
+#'
+#' @seealso
+#' \code{\link{FedInvestData}} the function to download the FedInvest data
+#'
+#' @importFrom YieldCurve Nelson.Siegel
+#' @importFrom YieldCurve NSrates
+#' @importFrom YieldCurve Svensson
+#' @importFrom YieldCurve Srates
+#'
+#' @examples
+#' library(ustreasuries)
+#' library(magrittr)
+#'
+#' start.date <- Sys.Date() - 25
+#' end.date   <- Sys.Date()
+#'
+#' rate_table_list <- dplyr::filter(FedInvestData(),
+#'                                   Date >= start.date &
+#'                                   Date <= end.date   &
+#'                                   SECURITY.TYPE != "TIPS") %>%
+#'     CoerceFedInvest_xts()
+#'
+#' rate_table_xts    <- rate_table_list$rate_table_xts
+#' unique_maturities <- rate_table_list$unique_maturities
+#' modified_data     <- rate_table_list$modified_data
+#'
+#' unique_maturities
+#'
+#' knitr::kable(as.data.frame(modified_data[30:32, c(2:4,10:14)]),
+#' caption="Sample of FedInvest Data with additional Data",
+#' row.names=FALSE)
+#'
+#' knitr::kable(as.data.frame(rate_table_xts[1:3, 1:5]),
+#' caption="Sample of FedInvest Data after Coersion to xts Format",
+#' row.names=TRUE)
+#'
+#' @export
+CoerceFedInvest_xts <- function(bond_data) {
+    # supresses
+    # Note: no visible binding for global variable ...
+    assign("Date", NULL)
+    assign("MATURITY.DATE", NULL)
+    assign("coupon", NULL)
+    assign("END.OF.DAY", NULL)
+    assign("ytm", NULL)
+
+    # find the fractional-year time to maturity for each bond
+    bond_data$fracyrs <- as.numeric((bond_data$MATURITY.DATE - bond_data$Date) / 365)
+
+    # the bonds have maturities all over the map, let's regularize them
+    # 1-month, 3-months, 6-months, years 1-30
+    # ===============================================================
+    mo1 = c(0,                   1/12+(3/12-1/12)/2)
+    mo3 = c(1/12+(3/12-1/12)/2,  3/12+(6/12-3/12)/2)
+    mo6 = c(3/12+(6/12-3/12)/2,  6/12+(12/12-6/12)/2)
+    yr1 = c(6/12+(12/12-6/12)/2, 1+(2-1)/2)
+    breaks = list(mo0.08333333=mo1, mo0.25=mo3, mo0.5=mo6, yr1=yr1)
+    for (i in seq(2,30)) {
+        breaks[[i+3]] <- c((i-1) + ((i)-(i-1))/2, (i) + ((i+1)-(i))/2)
+    }
+    names(breaks) <- c("mo0.08333333", "mo0.25", "mo0.5", "yr1", paste0("yr",seq(2,30)))
+
+    maturities = unlist(sapply(bond_data$fracyrs, function(x) {
+        for (i in 1:length(breaks)) {
+            if (x >= breaks[[i]][1] & x <= breaks[[i]][2]) {
+                return(as.numeric(substr(names(breaks[i]),3,nchar(names(breaks[i])))))
+            }
+        }
+    }))
+    bond_data$maturities <- maturities
+    # ===============================================================
+
+    # create fractional coupons from percent-format RATE
+    # ======================================================
+    matches <- regexpr("(\\d+\\.\\d+)%", bond_data$RATE, perl=TRUE);
+    result  <- attr(matches, "capture.start")[,1]
+    attr(result, "match.length") <- attr(matches, "capture.length")[,1]
+    bond_data$coupon <- as.numeric(regmatches(bond_data$RATE, result))/100
+
+    # find the mid-point between BUY & SELL
+    # =====================================
+    bond_data$mid <- apply(bond_data, 1, function(x) {
+                                            buy       <- as.numeric(x["BUY"])
+                                            sell      <- as.numeric(x["SELL"])
+                                            lower     <- min(buy, sell)
+                                            half_diff <- abs(buy - sell) / 2
+                                            mid       <- lower + half_diff
+                                            return(mid)
+                                        })
+
+    # find the YTM for each bond
+    # ==========================
+    bond_data$YTM <- apply(bond_data, 1, function(x) {
+                    Date   <- as.Date(x["Date"], origin="1970-01-01")
+                    mature <- as.Date(x["MATURITY.DATE"], origin="1970-01-01")
+                    coupon <- as.numeric(x["coupon"])
+
+                    # don't know which is better, mid or eod
+                    # but
+                    #     eod sometimes zero
+                    #     mid sometimes gives weird answers
+                    price  <- as.numeric(x["mid"])
+                    eod    <- as.numeric(x["END.OF.DAY"])
+                    price  <- eod
+
+                    if (grepl("BILL", x["SECURITY.TYPE"], perl=TRUE)) {
+                        tryCatch({
+                            jrvFinance::bond.yields(settle=Date,
+                                                    mature=mature,
+                                                    coupon=coupon,
+                                                    price=price,
+                                                    convention = "ACT/360")
+                            },
+                            error   = function(e) {NA},
+                            warning = function(e) {NA}
+                        )
+                    } else {
+                        tryCatch({
+                            jrvFinance::bond.yields(settle=Date,
+                                                    mature=mature,
+                                                    coupon=coupon,
+                                                    price=price,
+                                                    convention = "ACT/ACT")
+                            },
+                            error   = function(e) {NA},
+                            warning = function(e) {NA}
+                        )
+                    }
+                })
+
+    # for each set of bonds
+    #   on a date
+    #     with the same regularized maturity
+    # find the average YTM in percentage format
+    #   and create one row per date, one column per maturity
+    # ===========================================
+    rate_table <- bond_data %>%
+        dplyr::group_by(Date, maturities) %>%
+        dplyr::summarise(ytm=mean(YTM))   %>%
+        tidyr::spread(maturities, ytm)
+
+    # if Newton-Raphson and Bisection failed  to find IRR/YTM
+    # ... remove the rows
+    rate_table <- rate_table[complete.cases(rate_table),]
+
+    # convert to xts format
+    rate_table_xts <- xts::xts(rate_table[,-c(1)], order.by = rate_table$Date)
+
+    # get the vector of regularized maturities
+    unique_maturities <- as.numeric(names(rate_table_xts))
+
+    return(list(rate_table_xts    = rate_table_xts,
+                unique_maturities = unique_maturities,
+                modified_data     = bond_data))
+}
+
+#' Convert rates to discount factors Z(0, T)
+#'
+#' @description Returns the discount factors associated with Nelson Seigel fitted rates
+#'
+#' @details \eqn{Z(0, T) = e^{-r(0, T) * T}}
+#'
+#' @param NS_rates the value returned by the \code{NSrates()} or \code{Srates()}functions
+#' @return \code{data.frame} rownames are the dates,
+#' colnames are the time to maturity in character format preceded by "X"
+#'
+#' @author
+#' George Fisher
+#'
+#' @references
+#' Veronesi, P., \emph{Fixed Income Securities} p67
+#'
+#'
+#' @export
+NSzeros <- function(NS_rates) {
+    # function to generate Z(0, T) = exp(-r(0, T) * T)
+    # see p 67 of Veronesi
+    foo <- function(r) {
+        # extract T from the column name
+        T <- unname(sapply(names(r),
+                           function(x) unname(as.numeric(substr(x,2,nchar(x))))))
+
+        # NOTE: r(0, T) is divided by 100 '(r/100)'
+        #       because the data is in "percent" format
+        #       7.82 means 7.82%, 0.0782
+        return(exp(-(r/100) * T))
+    }
+    rates <- t(apply(NS_rates, 1, foo))
+    return(rates)
 }